+++ packages/backend/src/services/ProjectService/ProjectService.ts
@@ -21,9 +21,11 @@ import {
     DashboardAvailableFilters,
     DashboardBasicDetails,
     DashboardFilters,
+    DateGranularity,
     DbtProjectType,
     deepEqual,
     DefaultSupportedDbtVersion,
+    DimensionType,
     Explore,
     ExploreError,
     fieldId as getFieldId,
@@ -34,12 +36,14 @@ import {
     ForbiddenError,
     formatRows,
     getDashboardFilterRulesForTables,
+    getDateDimension,
     getDimensions,
     getFields,
     getItemId,
     getItemMap,
     getMetrics,
     hasIntersection,
+    isDateItem,
     isExploreError,
     isFilterableDimension,
     isUserWithOrg,
@@ -71,6 +75,7 @@ import {
     TableCalculationFormatType,
     TablesConfiguration,
     TableSelectionType,
+    TimeFrames,
     UpdateProject,
     UpdateProjectMember,
     UserAttributeValueMap,
@@ -92,6 +97,7 @@ import EmailClient from '../../clients/EmailClient/EmailClient';
 import { lightdashConfig } from '../../config/lightdashConfig';
 import { errorHandler } from '../../errors';
 import Logger from '../../logging/logger';
+import { AnalyticsModel } from '../../models/AnalyticsModel';
 import { JobModel } from '../../models/JobModel/JobModel';
 import { OnboardingModel } from '../../models/OnboardingModel/OnboardingModel';
 import { ProjectModel } from '../../models/ProjectModel/ProjectModel';
@@ -119,6 +125,7 @@ type RunQueryTags = {
     user_uuid?: string;
     organization_uuid?: string;
     chart_uuid?: string;
+    dashboard_uuid?: string;
 };
 
 type ProjectServiceDependencies = {
@@ -131,6 +138,7 @@ type ProjectServiceDependencies = {
     sshKeyPairModel: SshKeyPairModel;
     userAttributesModel: UserAttributesModel;
     s3CacheClient: S3CacheClient;
+    analyticsModel: AnalyticsModel;
 };
 
 export class ProjectService {
@@ -154,6 +162,8 @@ export class ProjectService {
 
     s3CacheClient: S3CacheClient;
 
+    analyticsModel: AnalyticsModel;
+
     constructor({
         projectModel,
         onboardingModel,
@@ -164,6 +174,7 @@ export class ProjectService {
         sshKeyPairModel,
         userAttributesModel,
         s3CacheClient,
+        analyticsModel,
     }: ProjectServiceDependencies) {
         this.projectModel = projectModel;
         this.onboardingModel = onboardingModel;
@@ -175,6 +186,7 @@ export class ProjectService {
         this.sshKeyPairModel = sshKeyPairModel;
         this.userAttributesModel = userAttributesModel;
         this.s3CacheClient = s3CacheClient;
+        this.analyticsModel = analyticsModel;
     }
 
     private async _resolveWarehouseClientSshKeys<
@@ -848,21 +860,93 @@ export class ProjectService {
         });
     }
 
+    static updateMetricQueryGranularity(
+        metricQuery: MetricQuery,
+        exploreDimensions: CompiledDimension[],
+        granularity?: DateGranularity,
+    ): {
+        metricQuery: MetricQuery;
+        oldDimension?: string;
+        newDimension?: string;
+    } {
+        if (granularity) {
+            const timeDimension = metricQuery.dimensions.find((dimension) => {
+                const dim = exploreDimensions.find(
+                    (d) => getFieldId(d) === dimension,
+                )?.type;
+                return (
+                    dim === DimensionType.TIMESTAMP ||
+                    dim === DimensionType.DATE
+                );
+            });
+
+            if (timeDimension) {
+                const { baseDimensionId } = getDateDimension(timeDimension);
+                if (!baseDimensionId) return { metricQuery };
+
+                const newTimeDimension = `${baseDimensionId}_${granularity.toLowerCase()}`;
+
+                // TODO replace sorts / filters ?
+                return {
+                    metricQuery: {
+                        ...metricQuery,
+                        dimensions: metricQuery.dimensions.map((dimension) =>
+                            dimension === timeDimension
+                                ? newTimeDimension
+                                : dimension,
+                        ),
+                        sorts: metricQuery.sorts.map((sort) =>
+                            sort.fieldId === timeDimension
+                                ? { ...sort, fieldId: newTimeDimension }
+                                : sort,
+                        ),
+                        tableCalculations: metricQuery.tableCalculations.map(
+                            (tc) => {
+                                const dim = exploreDimensions.find(
+                                    (d) => getFieldId(d) === timeDimension,
+                                );
+
+                                if (!dim) return tc;
+
+                                const baseDim = getDateDimension(dim.name);
+                                if (!baseDim) return tc;
+
+                                const oldDimension = `${dim.table}.${dim.name}`;
+                                // Rebuild the newDimension instead of looking at dimensions,
+                                // so we can even filter missing time frames from explore
+                                const newDimension = `${dim.table}.${
+                                    baseDim.baseDimensionId
+                                }_${granularity.toLowerCase()}`;
+
+                                return {
+                                    ...tc,
+                                    sql: tc.sql.replaceAll(
+                                        oldDimension,
+                                        newDimension,
+                                    ),
+                                };
+                            },
+                        ),
+                    },
+                    oldDimension: timeDimension,
+                    newDimension: newTimeDimension,
+                };
+            }
+        }
+        return { metricQuery };
+    }
+
     async runViewChartQuery({
         user,
         chartUuid,
-        dashboardFilters,
         versionUuid,
         invalidateCache,
-        dashboardSorts,
     }: {
         user: SessionUser;
         chartUuid: string;
-        dashboardFilters?: DashboardFilters;
         versionUuid?: string;
         invalidateCache?: boolean;
-        dashboardSorts?: SortField[];
-    }): Promise<ApiChartAndResults> {
+    }): Promise<ApiQueryResults> {
         if (!isUserWithOrg(user)) {
             throw new ForbiddenError('User is not part of an organization');
         }
@@ -903,29 +987,7 @@ export class ProjectService {
             throw new ForbiddenError();
         }
 
-        const tables = Object.keys(explore.tables);
-        const appliedDashboardFilters = dashboardFilters
-            ? {
-                  dimensions: getDashboardFilterRulesForTables(
-                      tables,
-                      dashboardFilters.dimensions,
-                  ),
-                  metrics: getDashboardFilterRulesForTables(
-                      tables,
-                      dashboardFilters.metrics,
-                  ),
-                  tableCalculations: getDashboardFilterRulesForTables(
-                      tables,
-                      dashboardFilters.tableCalculations,
-                  ),
-              }
-            : undefined;
-        const metricQuery: MetricQuery = appliedDashboardFilters
-            ? addDashboardFiltersToMetricQuery(
-                  savedChart.metricQuery,
-                  appliedDashboardFilters,
-              )
-            : savedChart.metricQuery;
+        const { metricQuery } = savedChart;
 
         const queryTags: RunQueryTags = {
             organization_uuid: organizationUuid,
@@ -934,34 +996,177 @@ export class ProjectService {
             chart_uuid: chartUuid,
         };
 
-        const metricWithOverrideSorting: MetricQuery = {
-            ...metricQuery,
-            sorts:
-                dashboardSorts && dashboardSorts.length > 0
-                    ? dashboardSorts
-                    : metricQuery.sorts,
-        };
-
         const { cacheMetadata, rows } = await this.runQueryAndFormatRows({
             user,
-            metricQuery: metricWithOverrideSorting,
+            metricQuery: savedChart.metricQuery,
             projectUuid,
             exploreName: savedChart.tableName,
             csvLimit: undefined,
-            context: dashboardFilters
-                ? QueryExecutionContext.DASHBOARD
-                : QueryExecutionContext.CHART,
+            context: QueryExecutionContext.CHART,
             queryTags,
             invalidateCache,
             explore,
         });
 
         return {
-            chart: savedChart,
-            explore,
             metricQuery,
             cacheMetadata,
             rows,
+        };
+    }
+
+    async getChartAndResults({
+        user,
+        chartUuid,
+        dashboardFilters,
+        invalidateCache,
+        dashboardSorts,
+        granularity,
+        dashboardUuid,
+    }: {
+        user: SessionUser;
+        chartUuid: string;
+        dashboardUuid: string;
+        dashboardFilters: DashboardFilters;
+        invalidateCache?: boolean;
+        dashboardSorts: SortField[];
+        granularity?: DateGranularity;
+    }): Promise<ApiChartAndResults> {
+        if (!isUserWithOrg(user)) {
+            throw new ForbiddenError('User is not part of an organization');
+        }
+
+        const savedChart = await this.savedChartModel.get(chartUuid);
+        const { organizationUuid, projectUuid } = savedChart;
+
+        if (
+            user.ability.cannot(
+                'view',
+                subject('SavedChart', { organizationUuid, projectUuid }),
+            ) ||
+            user.ability.cannot(
+                'view',
+                subject('Project', {
+                    organizationUuid,
+                    projectUuid,
+                }),
+            )
+        ) {
+            throw new ForbiddenError();
+        }
+
+        const [space, explore] = await Promise.all([
+            this.spaceModel.getSpaceSummary(savedChart.spaceUuid),
+            this.getExplore(
+                user,
+                projectUuid,
+                savedChart.tableName,
+                organizationUuid,
+            ),
+        ]);
+
+        if (!hasSpaceAccess(user, space)) {
+            throw new ForbiddenError();
+        }
+
+        await this.analyticsModel.addChartViewEvent(
+            savedChart.uuid,
+            user.userUuid,
+        );
+
+        const tables = Object.keys(explore.tables);
+        const appliedDashboardFilters = {
+            dimensions: getDashboardFilterRulesForTables(
+                tables,
+                dashboardFilters.dimensions,
+            ),
+            metrics: getDashboardFilterRulesForTables(
+                tables,
+                dashboardFilters.metrics,
+            ),
+            tableCalculations: getDashboardFilterRulesForTables(
+                tables,
+                dashboardFilters.tableCalculations,
+            ),
+        };
+        const metricQueryWithDashboardOverrides: MetricQuery = {
+            ...addDashboardFiltersToMetricQuery(
+                savedChart.metricQuery,
+                appliedDashboardFilters,
+            ),
+            sorts:
+                dashboardSorts && dashboardSorts.length > 0
+                    ? dashboardSorts
+                    : savedChart.metricQuery.sorts,
+        };
+
+        const queryTags: RunQueryTags = {
+            organization_uuid: organizationUuid,
+            project_uuid: projectUuid,
+            user_uuid: user.userUuid,
+            chart_uuid: chartUuid,
+            dashboard_uuid: dashboardUuid,
+        };
+
+        const exploreDimensions = getDimensions(explore);
+
+        const {
+            metricQuery: metricWithOverrideGranularity,
+            oldDimension,
+            newDimension,
+        } = ProjectService.updateMetricQueryGranularity(
+            metricQueryWithDashboardOverrides,
+            exploreDimensions,
+            granularity,
+        );
+
+        const { cacheMetadata, rows } = await this.runQueryAndFormatRows({
+            user,
+            metricQuery: metricWithOverrideGranularity,
+            projectUuid,
+            exploreName: savedChart.tableName,
+            csvLimit: undefined,
+            context: QueryExecutionContext.DASHBOARD,
+            queryTags,
+            invalidateCache,
+            explore,
+            granularity,
+        });
+
+        // TODO quick hack to get the old results on the chart with new granularity
+        // We should investigate if we can make the changes in the frontend
+        // to get the right field instead.
+        const rowsWithGranularity =
+            oldDimension && newDimension
+                ? rows.map((row) => ({
+                      ...row,
+                      [oldDimension]: row[newDimension],
+                  }))
+                : rows;
+        const metricQueryDimensions = [
+            ...metricQueryWithDashboardOverrides.dimensions,
+            ...(metricQueryWithDashboardOverrides.customDimensions ?? []),
+        ];
+        const hasADateDimension = exploreDimensions.find(
+            (c) =>
+                metricQueryDimensions.includes(getFieldId(c)) && isDateItem(c),
+        );
+
+        if (hasADateDimension) {
+            metricQueryWithDashboardOverrides.metadata = {
+                hasADateDimension: {
+                    name: hasADateDimension.name,
+                    label: hasADateDimension.label,
+                },
+            };
+        }
+
+        return {
+            chart: savedChart,
+            explore,
+            metricQuery: metricQueryWithDashboardOverrides,
+            cacheMetadata,
+            rows: rowsWithGranularity,
             appliedDashboardFilters,
         };
     }
@@ -1015,6 +1220,7 @@ export class ProjectService {
         queryTags,
         invalidateCache,
         explore,
+        granularity,
     }: {
         user: SessionUser;
         metricQuery: MetricQuery;
@@ -1025,6 +1231,7 @@ export class ProjectService {
         queryTags?: RunQueryTags;
         invalidateCache?: boolean;
         explore?: Explore;
+        granularity?: DateGranularity;
     }): Promise<ApiQueryResults> {
         return wrapOtelSpan(
             'ProjectService.runQueryAndFormatRows',
@@ -1040,6 +1247,7 @@ export class ProjectService {
                     queryTags,
                     invalidateCache,
                     explore,
+                    granularity,
                 });
                 span.setAttribute('rows', rows.length);
 
@@ -1251,6 +1459,7 @@ export class ProjectService {
         queryTags,
         invalidateCache,
         explore,
+        granularity,
     }: {
         user: SessionUser;
         metricQuery: MetricQuery;
@@ -1261,6 +1470,7 @@ export class ProjectService {
         queryTags?: RunQueryTags;
         invalidateCache?: boolean;
         explore?: Explore;
+        granularity?: DateGranularity;
     }): Promise<{ rows: Record<string, any>[]; cacheMetadata: CacheMetadata }> {
         const tracer = opentelemetry.trace.getTracer('default');
         return tracer.startActiveSpan(
@@ -1382,6 +1592,7 @@ export class ProjectService {
                             ).length,
                             context,
                             ...countCustomDimensionsInMetricQuery(metricQuery),
+                            dateZoomGranularity: granularity || null,
                         },
                     });