+++ packages/cli/src/handlers/compile.ts
@@ -49,7 +49,7 @@ export const compile = async (options: CompileHandlerOptions) => {
     if (!isSupportedDbtVersion(dbtVersion)) {
         if (process.env.CI === 'true') {
             console.error(
-                `Your dbt version ${dbtVersion} does not match our supported versions (1.3.* - 1.6.*), this could cause problems on compile or validation.`,
+                `Your dbt version ${dbtVersion} does not match our supported versions (1.3.* - 1.7.*), this could cause problems on compile or validation.`,
             );
         } else {
             const answers = await inquirer.prompt([
@@ -57,7 +57,7 @@ export const compile = async (options: CompileHandlerOptions) => {
                     type: 'confirm',
                     name: 'isConfirm',
                     message: `${styles.warning(
-                        `Your dbt version ${dbtVersion} does not match our supported version (1.3.* - 1.6.*), this could cause problems on compile or validation.`,
+                        `Your dbt version ${dbtVersion} does not match our supported version (1.3.* - 1.7.*), this could cause problems on compile or validation.`,
                     )}\nDo you still want to continue?`,
                 },
             ]);
@@ -66,7 +66,11 @@ export const compile = async (options: CompileHandlerOptions) => {
             }
         }
     }
-    await dbtCompile(options);
+
+    // Skipping assumes manifest.json already exists.
+    if (!options.skipDbtCompile) {
+        await dbtCompile(options);
+    }
 
     const absoluteProjectPath = path.resolve(options.projectDir);
     const absoluteProfilesPath = path.resolve(options.profilesDir);
@@ -146,7 +150,9 @@ ${errors.join('')}`),
         validModelsWithTypes,
         false,
         manifest.metadata.adapter_type,
-        manifestVersion === DbtManifestVersion.V10
+        [DbtManifestVersion.V10, DbtManifestVersion.V11].includes(
+            manifestVersion,
+        )
             ? []
             : Object.values(manifest.metrics),
         warehouseClient,