+++ packages/common/src/types/dbt.ts
@@ -67,7 +67,7 @@ type DbtModelJoin = {
     alias?: string;
     label?: string;
     type?: DbtModelJoinType;
-
+    hidden?: boolean;
     fields?: string[];
 };
 type DbtColumnMetadata = DbtColumnLightdashConfig & {};
@@ -420,4 +420,5 @@ export enum DbtManifestVersion {
     V8 = 'v8',
     V9 = 'v9',
     V10 = 'v10',
+    V11 = 'v11',
 }