+++ packages/frontend/src/hooks/dashboard/useDashboardChart.ts
@@ -1,17 +1,22 @@
 import { useDashboardContext } from '../../providers/DashboardProvider';
-import { useChartResults } from '../useQueryResults';
+import { useChartAndResults } from '../useQueryResults';
 import useDashboardFiltersForTile from './useDashboardFiltersForTile';
 
 const useDashboardChart = (tileUuid: string, savedChartUuid: string | null) => {
+    const dashboardUuid = useDashboardContext((c) => c.dashboard?.uuid);
     const invalidateCache = useDashboardContext((c) => c.invalidateCache);
     const dashboardFilters = useDashboardFiltersForTile(tileUuid);
     const chartSort = useDashboardContext((c) => c.chartSort);
     const tileSort = chartSort[tileUuid] || [];
-    return useChartResults(
+    const granularity = useDashboardContext((c) => c.dateZoomGranularity);
+
+    return useChartAndResults(
         savedChartUuid,
+        dashboardUuid ?? null,
         dashboardFilters,
         tileSort,
         invalidateCache,
+        granularity,
     );
 };