+++ packages/frontend/src/providers/DashboardProvider.tsx
@@ -7,6 +7,7 @@ import {
     Dashboard,
     DashboardFilterRule,
     DashboardFilters,
+    DateGranularity,
     fieldId,
     FilterableField,
     isDashboardChartTileType,
@@ -87,6 +88,14 @@ type DashboardContext = {
     hasChartTiles: boolean;
     chartSort: Record<string, SortField[]>;
     setChartSort: (sort: Record<string, SortField[]>) => void;
+    dateZoomGranularity: DateGranularity | undefined;
+    setDateZoomGranularity: Dispatch<
+        SetStateAction<DateGranularity | undefined>
+    >;
+    chartsWithDateZoomApplied: Set<string> | undefined;
+    setChartsWithDateZoomApplied: Dispatch<
+        SetStateAction<Set<string> | undefined>
+    >;
 };
 
 const Context = createContext<DashboardContext | undefined>(undefined);
@@ -109,7 +118,6 @@ export const DashboardProvider: React.FC<{
                     const overriddenDimensions = applyDimensionOverrides(
                         d.filters,
                         schedulerFilters,
-                        true,
                     );
 
                     return {
@@ -141,6 +149,27 @@ export const DashboardProvider: React.FC<{
     const [invalidateCache, setInvalidateCache] = useState<boolean>(false);
 
     const [chartSort, setChartSort] = useState<Record<string, SortField[]>>({});
+
+    const [dateZoomGranularity, setDateZoomGranularity] = useState<
+        DateGranularity | undefined
+    >(undefined);
+    const [chartsWithDateZoomApplied, setChartsWithDateZoomApplied] =
+        useState<Set<string>>();
+
+    // Update dashboard url date zoom change
+    useEffect(() => {
+        const newParams = new URLSearchParams(search);
+        if (dateZoomGranularity === undefined) {
+            newParams.delete('dateZoom');
+        } else {
+            newParams.set('dateZoom', dateZoomGranularity.toLowerCase());
+        }
+
+        history.replace({
+            pathname,
+            search: newParams.toString(),
+        });
+    }, [dateZoomGranularity, search, history, pathname]);
 
     const {
         overridesForSavedDashboardFilters,
@@ -252,9 +281,21 @@ export const DashboardProvider: React.FC<{
         }
     }, [dashboard?.filters, overridesForSavedDashboardFilters]);
 
-    // Gets filters from URL and storage after redirect
+    // Gets filters and dateZoom from URL and storage after redirect
     useMount(() => {
         const searchParams = new URLSearchParams(search);
+
+        // Date zoom
+        const dateZoomParam = searchParams.get('dateZoom');
+        if (dateZoomParam) {
+            const dateZoom = Object.values(DateGranularity).find(
+                (granularity) =>
+                    granularity.toLowerCase() === dateZoomParam?.toLowerCase(),
+            );
+            if (dateZoom) setDateZoomGranularity(dateZoom);
+        }
+
+        // Temp filters
         const tempFilterSearchParam = searchParams.get('tempFilters');
         const unsavedDashboardFiltersRaw = sessionStorage.getItem(
             'unsavedDashboardFilters',
@@ -522,6 +563,10 @@ export const DashboardProvider: React.FC<{
         hasChartTiles,
         chartSort,
         setChartSort,
+        dateZoomGranularity,
+        setDateZoomGranularity,
+        chartsWithDateZoomApplied,
+        setChartsWithDateZoomApplied,
     };
     return <Context.Provider value={value}>{children}</Context.Provider>;
 };